requirements.txt
flask
transformers
torch
sentencepiece
openai-whisper
python-multipart
ffmpeg-python
